+++ src/statebot.js
@@ -167,6 +167,7 @@ import {
   Logger,
   ReferenceCounter,
   Pausables,
+  Once,
 } from './utils'
 
 import {
@@ -304,6 +305,7 @@ function Statebot (name, options) {
     // (transition-only, and maybe a then-function too)
     const transitionConfigs = expandTransitions(transitionsOnly, canWarn)
 
+    // Now install handlers for on/performTransitions:
     const allCleanupFns =
       Object
         .entries(eventsMappedToTransitionConfigs)
@@ -342,7 +344,7 @@ function Statebot (name, options) {
       const route = `${fromState}->${toState}`
       return [
         routesHandled.increase(route),
-        onInternalEvent(route, action)
+        onInternalEvent(route, bindActionTo(toState, action))
       ]
     }
 
@@ -368,7 +370,7 @@ function Statebot (name, options) {
     function ifStateThenEnterState ({ fromState, toState, action, args }) {
       return inState(fromState, () => {
         enter(toState, ...args)
-        isFunction(action) && action(...args)
+        isFunction(action) && runActionFor(toState, action, ...args)
         return true
       })
     }
@@ -386,6 +388,22 @@ function Statebot (name, options) {
           }
         })
       ]
+    }
+
+    function runActionFor(state, actionFn, ...args) {
+      const onExitingState = actionFn(...args)
+      if (isFunction(onExitingState)) {
+        const uninstall = Once(enterExitMethods[ON_EXITING](state, (toState) => {
+          uninstall()
+          onExitingState(toState)
+        }))
+
+        allCleanupFns.push(uninstall)
+      }
+    }
+
+    function bindActionTo(state, actionFn) {
+      return (...args) => runActionFor(state, actionFn, ...args)
     }
   }
 
@@ -1456,6 +1474,11 @@ function Statebot (name, options) {
     /**
      * Run callbacks when transitions happen.
      *
+     * Since v2.8.0:
+     * - If a callback returns a function, it will be invoked when
+     *   the state is exited in the same manner as if an {@link #statebotfsmonexiting .onExiting()}
+     *   handler was created using it.
+     *
      * @memberof statebotFsm
      * @instance
      * @function
@@ -1555,6 +1578,11 @@ function Statebot (name, options) {
      * Perform transitions when events happen.
      *
      * Use `then` to optionally add callbacks to those transitions.
+     *
+     * Since v2.8.0:
+     * - If a `then` method returns a function, it will be invoked when
+     *   the state is exited in the same manner as if an {@link #statebotfsmonexiting .onExiting()}
+     *   handler was created using it.
      *
      * @memberof statebotFsm
      * @instance